Fixed Consultant report to show name of each consultant/user. Started working on Adding User's form/insert.
TODO: Need to correct mistake in Appointments where the User/Consultant is just the logged in user. This should be selectable.
Also need to work on a Report to select user and see all appointments by that user.
Then complete the Delete and modify users.
Finally need to go through and do data validation and error checking